requirements.txt
openai==1.2.4
python-dotenv == 1.0.0
tk